% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.r
\docType{methods}
\name{angles}
\alias{angles}
\title{Angles}
\usage{
angles(object, z = "thetaIn", interpolate = F, variogram.fit = T,
  plot.fit = F, aggregate = F, aggregate.fact = 100, ...)

\S4method{angles}{GeolocationPoints}(object, z = "thetaIn", interpolate = F,
  variogram.fit = T, plot.fit = F, aggregate = F, aggregate.fact = 100,
  ...)

\S4method{angles}{SAR}(object, z = "thetaIn", interpolate = F,
  variogram.fit = T, plot.fit = F, aggregate = F, aggregate.fact = 100,
  ...)

\S4method{angles}{SARSet}(object, z = "thetaIn", interpolate = F,
  variogram.fit = T, plot.fit = F, aggregate = F, aggregate.fact = 100,
  ...)
}
\arguments{
\item{object}{object of the \code{\link{SAR-class}} or a subclass (e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{z}{either incidence (default) or elevation angles 'thetaEl'.}

\item{interpolate}{TRUE if angles should be interpolated.}

\item{variogram.fit}{logical. Fit a Gaussian variogram?}

\item{plot.fit}{logical. Plot the fitted variogram?}

\item{aggregate}{logical. Should the raster be aggregated? 
If no aggregation is done the interpolation might take a very long time.}

\item{aggregate.fact}{integer. Aggregation factor expressed as number of cells in each direction (horizontally and vertically). 
Or two integers (horizontal and vertical aggregation factor) or three integers (when also aggregating over layers). 
See \code{\link{aggregate}} method of raster package.}
}
\description{
Returns angles provided by GCPs/TPs as a \code{\link{SpatialPointsDataFrame}} object.
}
\examples{
data(kili)
angles(master)
plot(angles(master))
}
\seealso{
\code{\link{plotAngles}}, \code{\link{GeolocationPoints-class}}
}

