% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.r
\docType{methods}
\name{crop,SAR-method}
\alias{crop,SAR-method}
\title{Crop}
\usage{
\S4method{crop}{SAR}(x, y, ...)
}
\arguments{
\item{x}{object object of the \code{\link{SAR-class}} or a subclass (e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{y}{Extent object, or any object from which an Extent object can be extracted.}

\item{...}{Additional arguments as for \code{\link[raster]{writeRaster}}.}
}
\description{
Crop a geographic subset (see \code{\link[raster]{crop}}).
}
\seealso{
\code{\link[raster]{crop}}
}

