% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparity.r
\name{disparity}
\alias{disparity}
\title{Disparity}
\usage{
disparity(h, theta1, theta2)
}
\arguments{
\item{h}{realtive height.}

\item{theta1}{angle.}

\item{theta2}{angle.}
}
\description{
Calculates the disparity \eqn{p} for given (incidence) angles.
\deqn{p \approx h  \cdot  (\cot  \theta_{i_2} - \cot \theta_{i_1})}
}
\examples{
plot((c(35:45)/360*2*pi), disparity(10, (c(35:45)/360*2*pi), 45/360*2*pi), type='l')
}
\seealso{
\code{\link{disparityMapExpected}}
}

