% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparityMapExpected.r
\name{disparityMapExpected}
\alias{disparityMapExpected}
\title{Expected (theoretical) disparity between two SAR records}
\usage{
disparityMapExpected(object, slave, h = 10, z = "thetaIn",
  variogram.fit = T, plot.fit = F, interpolate = T,
  aggregate.fact = 100, ...)
}
\arguments{
\item{object}{object of the \code{\link{SAR-class}} or a subclass (e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{slave}{object of the \code{\link{SAR-class}} or a subclass (e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{h}{numeric. Relative height.}

\item{z}{either incidence (default) or elevation angles 'thetaEl'.}

\item{variogram.fit}{TRUE if a Gaussian variogram should be fitted}

\item{plot.fit}{TRUE if the fitted variogram should be plotted}

\item{interpolate}{TRUE if angles should be interpolated.}

\item{aggregate.fact}{integer. Aggregation factor expressed as number of cells in each direction (horizontally and vertically). 
Or two integers (horizontal and vertical aggregation factor) or three integers (when also aggregating over layers). 
See \code{\link{aggregate}} method of raster package. If aggregate.fact equals 1, no aggregation is done, 
but then the interpolation might take a very long time...}
}
\description{
Returns the expected disparity within the intersection area of two SAR records.
Uses the angles provided by GCPs/TPs of both records. 
The angles for the slave image are interpolated using Kriging (\code{\link{krige}}).
Returns the results as a \code{\link{SpatialPointsDataFrame}} object.
}
\examples{
data(kili)
anglesDif(master, slave)
}
\seealso{
\code{\link{angles}}, \code{\link{plotAnglesDif}}, \code{\link{GeolocationPoints-class}}
}

