% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.r
\docType{class}
\name{GeolocationPoints-class}
\alias{GeolocationPoints-class}
\title{Geolocation points class}
\description{
A class to handle geolocation control points (GCPs) and tie points (TPs).
This class inherits from the \code{\link{SpatialPointsDataFrame}} class of the sp package.
}
\seealso{
\code{\link{SpatialPointsDataFrame}}, \code{\link{SAR-class}}
}

