% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolocationPoints.r
\docType{methods}
\name{geolocationPoints}
\alias{geolocationPoints}
\alias{geolocationPoints,SAR-method}
\alias{geolocationPoints,SARSet-method}
\title{Retrieve geolocation points}
\usage{
geolocationPoints(object, ...)

\S4method{geolocationPoints}{SAR}(object, ...)

\S4method{geolocationPoints}{SARSet}(object, ...)
}
\arguments{
\item{object}{object of the \code{\link{SAR-class}} or a subclass 
(e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}) or 
an object of the \code{\link{SARSet-class}}.}
}
\value{
either a single \code{\link{SpatialPointsDataFrame}} or a list of them
}
\description{
Returns GCPs/TPs as a \code{\link{SpatialPointsDataFrame}} object.
}
\examples{
data(kili)

geolocationPoints(master)
master@geolocationPoints
plot(geolocationPoints(master))

geolocationPoints(kili)
}
\seealso{
\code{\link{GeolocationPoints-class}}
}

