% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getAverageCpp}
\alias{getAverageCpp}
\title{Window mean written in C++}
\usage{
getAverageCpp(img, u, v, n)
}
\arguments{
\item{img}{Matrix.}

\item{u}{Integer. Center of window (column, i.e. y coordinate).}

\item{v}{Integer. Center of window (row, i.e. x coordinate).}

\item{n}{Integer. (Window size-1)/2 in x and y direction.}
}
\description{
Calculates window mean for a given \code{\link{SAR-class}} object.
The window is a square.
}

