% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightRel.r
\name{heightRel}
\alias{heightRel}
\title{Relative height}
\usage{
heightRel(p, theta1, theta2)
}
\arguments{
\item{p}{disparity height.}

\item{theta1}{angle.}

\item{theta2}{angle.}
}
\description{
Calculates the relative height \eqn{h} for a given disparity \eqn{p} and (incidence) angles.
\deqn{h \approx \frac{p}{\cot  \theta_{i_2} - \cot \theta_{i_1}}}
}
\examples{
plot((c(35:45)/360*2*pi), disparity(10, (c(35:45)/360*2*pi), 45/360*2*pi), type='l')
}
\seealso{
\code{\link{disparityMapExpected}}
}

