% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnglesDif.r
\docType{methods}
\name{plotAnglesDif}
\alias{plotAnglesDif}
\title{Plot differences of two SAR records}
\usage{
plotAnglesDif(object, slave, z = "thetaIn", xlab = "Longitude",
  ylab = "Latitude", breaks = 255,
  col.regions = colorRampPalette(colors)(breaks), cex = 1, cex.axis = 1,
  grid = T, variogram.fit = T, plot.fit = F, plot.legend = T,
  pch = 20, legend.lab = expression(paste(Delta, theta[i], " [deg]")),
  interpolate = F, aggregate.fact = 100, disparity = F, h = 100, ...)

\S4method{plotAnglesDif}{SAR,SAR}(object, slave, z, xlab, ylab, breaks,
  col.regions, cex, cex.axis, grid, variogram.fit, plot.fit, plot.legend, pch,
  legend.lab, interpolate, aggregate.fact, disparity, h,
  xlim = c(min(object@extent@xmin, slave@extent@xmin), max(object@extent@xmax,
  slave@extent@xmax)), ylim = c(min(object@extent@ymin, slave@extent@ymin),
  max(object@extent@ymax, slave@extent@ymax)), asp = 1/cos((mean(range(ylim))
  * pi)/180), zlim = c(min(angleDif@data[, 1]), max(angleDif@data[, 1])), ...)

\S4method{plotAnglesDif}{SARSet,ANY}(object, slave, z = "thetaIn",
  xlab = "Longitude", ylab = "Latitude", breaks = 255,
  col.regions = colorRampPalette(colors)(breaks), cex = 1, cex.axis = 1,
  grid = T, variogram.fit = T, plot.fit = F, plot.legend = T,
  pch = 20, legend.lab = expression(paste(Delta, theta[i], " [deg]")),
  interpolate = F, aggregate.fact = 100, disparity = F, h = 100, ...)
}
\arguments{
\item{object}{Object of the \code{\link{SAR-class}} or a subclass (e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{z}{Character. Either incidence ('thetaIn', default) or elevation angles 'thetaEl'.}

\item{xlab}{Character. Title of the x axis.}

\item{ylab}{Character. Title of the y axis.}

\item{breaks}{Integer. Number of breaks of the color ramp.}

\item{col.regions}{Color ramp.}

\item{variogram.fit}{Logical. Fit a Gaussian variogram?}

\item{plot.legend}{Logical. Plot the legend?}

\item{legend.lab}{Character. Title of the legend.}

\item{interpolate}{Logical. Interpolate angles?}

\item{aggregate.fact}{Integer. Aggregation factor expressed as number of cells in each direction (horizontally and vertically). 
Or two integers (horizontal and vertical aggregation factor) or three integers (when also aggregating over layers). 
See \code{\link{aggregate}} method of raster package. If no aggregation is done, 
the interpolation might take a very long time...}

\item{disparity}{Logical. Should the expected disparity be plotted?}

\item{h}{Numeric. Relative height for plotting the expected disparity.}

\item{...}{graphical parameters. Any argument that can be passed to plot, such as axes=FALSE and main='title'.}

\item{asp}{Numeric vector. Aspect, default is 1/cos((mean(range(ylim)) * pi)/180).}

\item{plot.fit.}{Logical. Plot the fitted variogram?}

\item{asp}{Numeric vector. Aspect, default is 1/cos((mean(range(ylim)) * pi)/180).}
}
\description{
Plots the difference in angles of two provided by GCPs/TPs as points.
}
\examples{
data(kili)

plotAnglesDif(kili[[1]], kili[[3]])

plotAnglesDif(kili)

plotAnglesDif(master, slave, interpolate=T)
plot(anglesDif(master, slave, interpolate=T))

plotAnglesDif(kili[[1]], kili[[5]], disparity = T,
legend.lab = 'Disparity [m]', h=100)
}
\seealso{
\code{\link{anglesDif}}, \code{\link{angles}}, \code{\link{plotAngles}}, \code{\link{GeolocationPoints-class}}
}

