% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDisparityMapExpected.r
\docType{methods}
\name{plotDisparityMapExpected}
\alias{plotDisparityMapExpected}
\title{Plot disparity map of two SAR records}
\usage{
plotDisparityMapExpected(object, slave, z = "thetaIn", xlab = "Longitude",
  ylab = "Latitude", h = 1, breaks = 255,
  col.regions = colorRampPalette(colors)(breaks), cex = 1, cex.axis = 1,
  grid = T, variogram.fit = T, plot.fit = F, plot.legend = T,
  pch = 20, legend.lab = "disparity [m]", interpolate = F,
  aggregate.fact = 100, ...)

\S4method{plotDisparityMapExpected}{SAR,SAR}(object, slave, z = "thetaIn",
  xlab = "Longitude", ylab = "Latitude", h = 1, breaks = 255,
  col.regions = colorRampPalette(colors)(breaks), cex = 1, cex.axis = 1,
  grid = T, variogram.fit = T, plot.fit = F, plot.legend = T,
  pch = 20, legend.lab = "disparity [m]", interpolate = F,
  aggregate.fact = 100, ...)

\S4method{plotDisparityMapExpected}{SARSet,ANY}(object, slave, z = "thetaIn",
  xlab = "Longitude", ylab = "Latitude", h = 1, breaks = 255,
  col.regions = colorRampPalette(colors)(breaks), cex = 1, cex.axis = 1,
  grid = T, variogram.fit = T, plot.fit = F, plot.legend = T,
  pch = 20, legend.lab = "disparity [m]", interpolate = F,
  aggregate.fact = 100, ...)
}
\arguments{
\item{object}{object of the \code{\link{SAR-class}} or a subclass (e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{z}{either incidence (default) or elevation angles 'thetaEl'.}

\item{xlab}{title of the x axis.}

\item{ylab}{title of the y axis.}

\item{breaks}{number of breaks of the color ramp.}

\item{col.regions}{color ramp.}

\item{variogram.fit}{TRUE if a Gaussian variogram should be fitted.}

\item{plot.fit}{TRUE if the fitted variogram should be plotted.}

\item{plot.legend}{TRUE if the legend is plotted.}

\item{legend.lab}{title of the legend.}

\item{...}{graphical parameters. Any argument that can be passed to plot, such as axes=FALSE and main='title'.}

\item{asp}{aspect, default is 1/cos((mean(range(ylim)) * pi)/180).}
}
\description{
Plots the expected disparity map of two SAR records. 
Either the expected disparity for a relative height \eqn{h} or the calculated disparity is used.
}
\examples{
data(kili)

plotDisparityMapExpected(kili[[1]], kili[[2]])
plot(disparityMapExpected(kili[[1]], kili[[2]]))

plotDisparityMapExpected(kili)
}
\seealso{
\code{\link{disparity}}, \code{\link{angles}}, \code{\link{GeolocationPoints-class}}
}

