% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.r
\name{plotMap}
\alias{plotMap}
\title{Plot map}
\usage{
plotMap(object, database = "world", regions = ".", xlim = c(-8, 8) +
  mean(object@extent[1:2]), ylim = c(-8, 8) + mean(object@extent[3:4]),
  xlab = "Longitude", ylab = "Latitude", main = NULL, add = F,
  map.default.text = T, sar = T, labels.x = NULL, labels.y = NULL,
  labels = seq_along(labels.x), col = rgb(1, 0, 0, 0.5), orbit = T, ...)
}
\arguments{
\item{object}{object of the \code{\link{SARSet-class}} 
or object of the \code{\link{SAR-class}} (or a subclass e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{database}{character string naming a geographical database (see \code{\link[maps]{map}}).}

\item{regions}{character vector that names the polygons to draw (see \code{\link[maps]{map}}).}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{ylim}{the y limits (y1, y2) of the plot.}

\item{xlab}{title of the x axis.}

\item{ylab}{title of the y axis.}

\item{map.default.text}{TRUE if default text is plotted using the \code{\link[maps]{map.text}} method.}

\item{sar}{TRUE if location should be plotted.}

\item{labels.x}{x coordinates of labels.}

\item{labels.y}{y coordinates of labels.}

\item{labels}{vector of text that should be plotted on the map.}

\item{col}{color within the recorded area.}

\item{orbit}{TRUE if orbit number is plotted}

\item{...}{graphical parameters of \code{\link[maps]{map}} method.}

\item{border}{color of the \code{\link{border}} of the SAR location.}
}
\description{
Plots a map that shows the location of SAR records.
Uses the \code{\link[maps]{map}} method of the maps package.
}
\examples{
data(kili)

plotMap(kili)
plotMap(kili[[1]])
}
\seealso{
\code{\link{plotGmap}}, \code{\link[maps]{map}}, \code{\link{border}}, \code{\link{plotBorder}}
}

