% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOrbitNumber.r
\name{plotOrbitNumber}
\alias{plotOrbitNumber}
\title{Plot orbit number}
\usage{
plotOrbitNumber(object, cex = 1, col = "black", panel = F, ...)
}
\arguments{
\item{object}{object of the \code{\link{SARSet-class}} 
or object of the \code{\link{SAR-class}} (or a subclass e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{panel}{TRUE if text is added to a levelplot.}

\item{...}{parameters of \code{\link{text}} method.}

\item{labels}{character that is ploted (default: object@orbit).}

\item{x}{x coordinates of labels.}

\item{y}{y coordinates of labels.}
}
\description{
Adds orbit numbers of SAR records.
}
\examples{
data(kili)

plotMap(kili[[1]], orbit=F)
plotOrbitNumber(kili[[1]])

plotMap(kili, orbit=F)
plotOrbitNumber(kili)
}
\seealso{
\code{\link{plotMap}}, \code{\link{plotGmap}}
}

