% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\docType{methods}
\name{plotSAR}
\alias{plot,SARSet,ANY-method}
\alias{plotSAR}
\alias{plotSAR,RasterLayer,ANY-method}
\alias{plotSAR,SAR,ANY-method}
\alias{plotSAR,SAR,SAR-method}
\alias{plotSAR,SARSet,ANY-method}
\title{Plot SAR record}
\usage{
plotSAR(object, slave = NULL, xlab = "Longitude", ylab = "Latitude",
  maxpixels = 50000, max.val = 500, cex.axis = 1, cex.lab = 1,
  legend.by = 30, plotIntersect = T, border = "red", grid = T,
  plot.legend = T, legend.lab = "Amplitude", legend.line = 2.2,
  breaks = seq(0, max.val, by = 1), legend.breaks = seq(0, max.val, by =
  legend.by), zlim = c(0, max.val), cols = colorRampPalette(c("grey0",
  "grey100"))(length(seq(0, max.val, by = 1)) - 1), north.arrow = F,
  north.arrow.pos = NULL, north.arrow.length = 0.05, scalebar = F,
  scalebar.pos = NULL, scalebar.col = "black", scalebar.d = 100, ...)

\S4method{plotSAR}{SAR,ANY}(object, slave = NULL, xlab, ylab, maxpixels,
  max.val, cex.axis, cex.lab, legend.by, plotIntersect, border, grid,
  plot.legend, legend.lab, legend.line, breaks, legend.breaks, zlim, cols,
  xlim = c(object@extent@xmin, object@extent@xmax),
  ylim = c(object@extent@ymin, object@extent@ymax), asp = NULL, ...)

\S4method{plotSAR}{RasterLayer,ANY}(object, slave = NULL,
  xlab = "Longitude", ylab = "Latitude", maxpixels = 50000,
  max.val = 500, cex.axis = 1, cex.lab = 1, legend.by = 30,
  plotIntersect = T, border = "red", grid = T, plot.legend = T,
  legend.lab = "Amplitude", legend.line = 2.2, breaks = seq(0, max.val, by
  = 1), legend.breaks = seq(0, max.val, by = legend.by), zlim = c(0,
  max.val), cols = colorRampPalette(c("grey0", "grey100"))(length(seq(0,
  max.val, by = 1)) - 1), north.arrow = F, north.arrow.pos = NULL,
  north.arrow.length = 0.05, scalebar = F, scalebar.pos = NULL,
  scalebar.col = "black", scalebar.d = 100, ...)

\S4method{plotSAR}{SAR,SAR}(object, slave, xlab, ylab, maxpixels, max.val,
  cex.axis, cex.lab, legend.by, plotIntersect, border, grid, plot.legend,
  legend.lab, legend.line, breaks, legend.breaks, zlim, cols,
  xlim = c(min(object@extent@xmin, slave@extent@xmin), max(object@extent@xmax,
  slave@extent@xmax)), ylim = c(min(object@extent@ymin, slave@extent@ymin),
  max(object@extent@ymax, slave@extent@ymax)), asp = NULL, ...)

\S4method{plotSAR}{SARSet,ANY}(object, slave = NULL, xlab = "Longitude",
  ylab = "Latitude", maxpixels = 50000, max.val = 500, cex.axis = 1,
  cex.lab = 1, legend.by = 30, plotIntersect = T, border = "red",
  grid = T, plot.legend = T, legend.lab = "Amplitude",
  legend.line = 2.2, breaks = seq(0, max.val, by = 1),
  legend.breaks = seq(0, max.val, by = legend.by), zlim = c(0, max.val),
  cols = colorRampPalette(c("grey0", "grey100"))(length(seq(0, max.val, by =
  1)) - 1), north.arrow = F, north.arrow.pos = NULL,
  north.arrow.length = 0.05, scalebar = F, scalebar.pos = NULL,
  scalebar.col = "black", scalebar.d = 100, ...)

\S4method{plot}{SARSet,ANY}(x, y = NULL, ...)
}
\arguments{
\item{object}{object of the \code{\link{SAR-class}} or a subclass 
(e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}) (or 
an object of the \code{\link{SARSet-class}}, then slave is ignored).}

\item{slave}{object of the \code{\link{SAR-class}} or a subclass 
(e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{xlab}{title of the x axis.}

\item{ylab}{title of the y axis.}

\item{maxpixels}{Maximum number of cells to use for the plot (see \code{\link{plot}} for RasterLayer object).}

\item{max.val}{maximal amplitude to plot, higher amplitudes are ploted in white. 
To use even the highest amplitudes type 'object@data@max'.}

\item{plotIntersect}{logical. Plot the intersection area?}

\item{border}{color of the border.}

\item{grid}{logical. Add a grid to the plot?}

\item{north.arrow}{logical. Plot a north arrow?}

\item{north.arrow.pos}{vector c(x,y) that contains the position of the north arrow.}

\item{north.arrow.length}{length of the north arrow.}

\item{scalebar}{logical. Plot a scalebar?}

\item{scalebar.pos}{vector c(x,y) that contains the position of the scalebar.}

\item{scalebar.col}{color of the scalebar.}

\item{scalebar.d}{distance in kilometers.}

\item{...}{plotting parameters like xlim, ylim and asp.}

\item{col}{color ramp (see \code{\link{colorRampPalette}}).}
}
\description{
Plot one SAR or two SAR records together.
}
\examples{
data(kili)
plotSAR(master)
plotSAR(slave) 
}

