% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.r
\docType{class}
\name{SARSet-class}
\alias{SARSet-class}
\title{SAR set class}
\description{
A class to handle a collection of synthetic aperture radar (SAR) records.
Objects of this class are lists.
}
\section{Slots}{

\describe{
\item{\code{crs}}{\code{\link{CRS}} object}
}}
\examples{
kiliSetAsc <- new('SARSet',
              c(kili_2016_06_05_155520_vv_warped,  # 130
              kili_2016_05_31_154724_vv_warped,    #  57
              kili_2015_04_19_154656_vv_warped))   #  57 
alternatively
sarSet(c(kili_2016_06_05_155520_vv_warped,         # 130
              kili_2016_05_31_154724_vv_warped,    #  57
              kili_2015_04_19_154656_vv_warped))   #  57
}
\seealso{
\code{\link{SAR-class}}
}

