% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarSet.r
\name{sarSet}
\alias{sarSet}
\title{Create a set of synthetic aperture radar (SAR) images}
\usage{
sarSet(elements = c(), ...)
}
\arguments{
\item{elements}{list of SAR objects.}
}
\description{
Creates a set of synthetic aperture radar (SAR) images. 
This is useful to identify stereo images.
}
\examples{
 
data(kili)
my.sar.set <- sarSet(c(kili[[1]], kili[[2]]))
}

