% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.r
\docType{class}
\name{Sentinel-class}
\alias{Sentinel-class}
\title{Sentinel class}
\description{
A class to handle Sentinel-1 SAR data. 
This class inherits from the \code{\link{SAR-class}}.
}
\section{Slots}{

\describe{
\item{\code{address}}{address of the SAR data main folder.}

\item{\code{polarization}}{polarization of the SAR record (e.g. VV or HH).}

\item{\code{geolocationPoints}}{\code{\link{GeolocationPoints-class}} object that contains GCPs/TPs.}

\item{\code{node}}{direction of the satellite (ascending or descending).}

\item{\code{orbit}}{orbit number (e.g. sentinel has 175 in total).}

\item{\code{imgSubstring}}{character. This substring is searched for when image data is read.}
}}
\examples{
# orbit 130, dual polarization vv and vh
kili_2014_12_25_155516_vv <- new('Sentinel', warped=F, polarization='vv',
                             address='sentinel1/kili/S1A_IW_GRDH_1SDV_20141225T155516_20141225T155541_003877_004A5A_2263.SAFE/')

# alternatively
sarRecord(address='sentinel1/kili/S1A_IW_GRDH_1SDV_20141225T155516_20141225T155541_003877_004A5A_2263.SAFE/',
          satellite='sentinel-1', warped=F, polarization='vv')

kili_2014_12_25_155516_vv@orbit
kili_2014_12_25_155516_vv@polarization
kili_2014_12_25_155516_vv@node
}
\seealso{
\code{\link{SAR-class}}, \code{\link{GeolocationPoints-class}}, \code{\link{SARSet-class}}
}

