% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zncc.r
\name{ssd}
\alias{ssd}
\title{Sum of squared differences (SSD)}
\usage{
ssd(master, slave, u1, v1, u2, v2, n = 3)
}
\arguments{
\item{master}{\code{\link{SAR-class}} object (or a subclass e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{slave}{\code{\link{SAR-class}} object (or a subclass e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{u1}{Integer. Column, i.e. y coordinate of master pixel.}

\item{v1}{Integer. Row, i.e. x coordinate of master pixel.}

\item{u2}{Integer. Column, i.e. y coordinate of slave pixel.}

\item{v2}{Integer. Row, i.e. x coordinate of slave pixel.}

\item{n}{Integer. (Window size-1)/2 in x and y direction.}
}
\description{
Calculates the sum of squared differences (SSD) between two \code{\link{SAR-class}} objects.
This method also works for \code{\link{RasterLayer-class}} objects. The minimum SSD value determines disparity.
}

