% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{znccCpp}
\alias{znccCpp}
\title{Zero normalized cross correlation (ZNCC)}
\usage{
znccCpp(master, slave, u1, v1, u2, v2, n)
}
\description{
Calculates the zero normalized cross correlation (ZNCC) coefficient between two \code{\link{SAR-class}} objects.
This method also works for \code{\link{RasterLayer-class}} objects. The maximum ZNCC value determines disparity.
}

