% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.r
\docType{methods}
\name{angles}
\alias{angles}
\title{Angles}
\usage{
angles(object, z = "thetaIn", interpolate = F, variogram.fit = T,
  plot.fit = F, aggregate = T, aggregate.fact = 1000, ...)

\S4method{angles}{GeolocationPoints}(object, z = "thetaIn", interpolate = F,
  variogram.fit = T, plot.fit = F, aggregate = T, aggregate.fact = 1000,
  ...)

\S4method{angles}{SAR}(object, z = "thetaIn", interpolate = F,
  variogram.fit = T, plot.fit = F, aggregate = T, aggregate.fact = 1000,
  ...)

\S4method{angles}{SARSet}(object, z = "thetaIn", interpolate = F,
  variogram.fit = T, plot.fit = F, aggregate = T, aggregate.fact = 1000,
  ...)
}
\arguments{
\item{object}{Object of the \code{\link{SAR-class}} or a subclass (e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{z}{Character. either incidence ('thetaIn', default) or elevation angles 'thetaEl'.}

\item{interpolate}{Logical. Interpolate angles?}

\item{variogram.fit}{Logical. Fit a Gaussian variogram?}

\item{plot.fit}{Logical. Plot the fitted variogram?}

\item{aggregate.fact}{Integer. Aggregation factor expressed as number of cells in each direction (horizontally and vertically). 
Or two integers (horizontal and vertical aggregation factor) or three integers (when also aggregating over layers). 
See \code{\link{aggregate}} method of raster package.}

\item{aggregate.}{Logical. Should the raster be aggregated? If no aggregation is done the interpolation might take a very long time.}
}
\value{
\code{\link{SpatialPointsDataFrame}}
}
\description{
Returns angles provided by GCPs/TPs as a \code{\link{SpatialPointsDataFrame}} object.
}
\examples{
data(kili)

angles(kili[[1]])
plotAngles(kili[[1]])
plotAngles(kili[[5]])

plot(angles(kili[[1]]))
spplot(angles(kili[[1]]))

plotAngles(master, interpolate = T, aggregate.fact=100)

angles.sp <- angles(master, interpolate = T, aggregate.fact=10)
angles.raster <- rasterFromXYZ(angles.sp)
plot(angles.raster)
}
\seealso{
\code{\link{plotAngles}}, \code{\link{GeolocationPoints-class}}
}

