% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglesDif.r
\name{anglesDif}
\alias{anglesDif}
\title{Angles differences of two SAR records}
\usage{
anglesDif(object, slave = NULL, z = "thetaIn", variogram.fit = T,
  plot.fit = F, interpolate = F, aggregate.fact = 1000, ...)
}
\arguments{
\item{object}{Object of the \code{\link{SAR-class}} or a subclass (e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{slave}{Object of the \code{\link{SAR-class}} or a subclass (e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{z}{Character. Either incidence ('thetaIn', default) or elevation angles 'thetaEl'.}

\item{variogram.fit}{Logical. Fit a Gaussian variogram?}

\item{plot.fit}{Logical. Plot the fitted variogram?}

\item{interpolate}{Logical. Interpolate angles?}

\item{aggregate.fact}{Integer. Aggregation factor expressed as number of cells in each direction (horizontally and vertically). 
Or two integers (horizontal and vertical aggregation factor) or three integers (when also aggregating over layers). 
See \code{\link{aggregate}} method of raster package.}
}
\value{
\code{\link{SpatialPointsDataFrame}}
}
\description{
Returns the differences in angles within the intersection area of two SAR records.
Uses the angles provided by GCPs/TPs of both records. 
The angles for the slave image are interpolated using Kriging (\code{\link{krige}}).
Returns the results as a \code{\link{SpatialPointsDataFrame}} object.
}
\examples{
data(kili)

anglesDif(kili[[1]], kili[[5]])
plotAnglesDif(kili[[1]], kili[[5]])

anglesDif(master, slave, interpolate=TRUE, aggregate.fact=100)
plotAnglesDif(master, slave, interpolate=TRUE, aggregate.fact=100)

anglesDif.sp <- anglesDif(master, slave, interpolate = TRUE, aggregate.fact=10)
anglesDif.raster <- rasterFromXYZ(anglesDif.sp)
plot(anglesDif.raster)
}
\seealso{
\code{\link{angles}}, \code{\link{plotAnglesDif}}, \code{\link{GeolocationPoints-class}}
}

