% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparityMapExpected.r
\name{disparityMapExpected}
\alias{disparityMapExpected}
\title{Expected (theoretical) disparity between two SAR records}
\usage{
disparityMapExpected(object, slave, h = 1, z = "thetaIn",
  variogram.fit = T, plot.fit = F, interpolate = F,
  aggregate.fact = 1000, ...)
}
\arguments{
\item{object}{Object of the \code{\link{SAR-class}} or a subclass (e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{slave}{Object of the \code{\link{SAR-class}} or a subclass (e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{h}{Numeric. Relative height.}

\item{z}{Character. Either incidence ('thetaIn', default) or elevation angles 'thetaEl'.}

\item{variogram.fit}{Logical. Fit a Gaussian variogram?}

\item{plot.fit}{Logical. Plot the fitted variogram?}

\item{interpolate}{Logical. Interpolate angles?}

\item{aggregate.fact}{Integer. Aggregation factor expressed as number of cells in each direction (horizontally and vertically). 
Or two integers (horizontal and vertical aggregation factor) or three integers (when also aggregating over layers). 
See \code{\link{aggregate}} method of raster package.}
}
\description{
Returns the expected disparity within the intersection area of two SAR records.
Uses the angles provided by GCPs/TPs of both records. 
The angles for the slave image are interpolated using Kriging (\code{\link{krige}}).
Returns the results as a \code{\link{SpatialPointsDataFrame}} object.
}
\examples{
data(kili)

disparityMapExpected(kili[[1]], kili[[5]], h=1)

disparityMapExpected(master, slave, interpolate = TRUE, aggregate.fact = 100)
}
\seealso{
\code{\link{disparityMap}}, \code{\link{angles}}, \code{\link{plotAnglesDif}}, \code{\link{GeolocationPoints-class}}
}

