% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zncc.r
\name{getAverage}
\alias{getAverage}
\title{Window mean}
\usage{
getAverage(object, u, v, n)
}
\arguments{
\item{object}{\code{\link{SAR-class}} object (or a subclass e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{u}{Integer. Center of window (column, i.e. y coordinate).}

\item{v}{Integer. Center of window (row, i.e. x coordinate).}

\item{n}{Integer. (Window size-1)/2 in x and y direction.}
}
\description{
Calculates window mean for a given \code{\link{SAR-class}} object.
The window is a square.
}

