% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBorder.r
\name{plotBorder}
\alias{plotBorder}
\title{Plot border}
\usage{
plotBorder(object, col = rgb(1, 0, 0, 0.5), panel = F, ...)
}
\arguments{
\item{object}{object of the \code{\link{SAR-class}} or a subclass (e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{col}{color within the recorded area.}

\item{panel}{TRUE if the border should be plotted on a lattice plot (e.g. levelplot).}
}
\description{
Plots border provided by metadata as a \code{\link{SpatialPolygons}} object.
}
\examples{
data(kili)

plotBorder(master)

plotMap(kili, sar=FALSE, orbit=FALSE)

plotBorder(kili[[1]], add=TRUE)
plotOrbitNumber(kili[[1]])
plotBorder(kili[[2]], add=TRUE)
plotOrbitNumber(kili[[2]])
}
\seealso{
\code{\link{border}}
}

