% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGmap.r
\name{plotGmap}
\alias{plotGmap}
\title{Plot google earth map}
\usage{
plotGmap(object, xlim = c(-5, 5) + mean(object@extent[1:2]), ylim = c(-5, 5)
  + mean(object@extent[3:4]), xlab = "Longitude", ylab = "Latitude",
  sar = T, col = rgb(1, 0, 0, 0.5), scale = 1, border = "black", ...)
}
\arguments{
\item{object}{object of the \code{\link{SARSet-class}} 
or object of the \code{\link{SAR-class}} (or a subclass e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{ylim}{the y limits (y1, y2) of the plot.}

\item{xlab}{title of the x axis.}

\item{ylab}{title of the y axis.}

\item{sar}{TRUE if location should be plotted.}

\item{col}{color within the recorded area.}

\item{border}{color of the \code{\link{border}} of the SAR location.}

\item{...}{graphical parameters of \code{\link{gmap}} method.}

\item{asp}{aspect, default is 1/cos((mean(range(ylim)) * pi)/180).}
}
\description{
Plots a google earth map that shows the location of SAR records. 
Uses the \code{\link{gmap}} method of the dismo package.
}
\examples{
plotGmap(master)
plotGmap(kili)
}
\seealso{
\code{\link{plotMap}}, \code{\link{gmap}}, \code{\link{border}}, \code{\link{plotBorder}}
}

