% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOrbitNumber.r
\docType{methods}
\name{plotOrbitNumber}
\alias{plotOrbitNumber}
\alias{plotOrbitNumber,SAR-method}
\alias{plotOrbitNumber,SARSet-method}
\title{Plot orbit number}
\usage{
plotOrbitNumber(object, cex = 1, col = "black", panel = F, ...)

\S4method{plotOrbitNumber}{SAR}(object, cex = cex, col = col, panel,
  x = object@centerLon, y = object@centerLat, labels = object@orbit, ...)

\S4method{plotOrbitNumber}{SARSet}(object, cex = 1, col = "black",
  panel = F, ...)
}
\arguments{
\item{object}{object of the \code{\link{SARSet-class}} 
or object of the \code{\link{SAR-class}} (or a subclass e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{panel}{TRUE if text is added to a levelplot.}

\item{...}{parameters of \code{\link{text}} method.}

\item{x}{x coordinates of labels.}

\item{y}{y coordinates of labels.}

\item{labels}{character that is ploted (default: object@orbit).}
}
\description{
Adds orbit numbers of SAR records.
}
\examples{
data(kili)

plotMap(kili[[1]], orbit=FALSE)
plotOrbitNumber(kili[[1]])

plotMap(kili, orbit=FALSE)
plotOrbitNumber(kili)
}
\seealso{
\code{\link{plotMap}}, \code{\link{plotGmap}}
}

