% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarProject.r
\docType{methods}
\name{sarProject}
\alias{sarProject}
\alias{sarProject,SAR-method}
\title{Project SAR object}
\usage{
sarProject(object, suffix = "-latlon",
  dstfile = paste0(sub("\\\\.[[:alnum:]]+$", "",
  as.character(object@file@name)), suffix, ".tiff"), of = "GTiff",
  r = "bilinear", t_srs = "+init=epsg:4326", dstalpha = T,
  dstnodata = "NA", overwrite = F, ...)

\S4method{sarProject}{SAR}(object, suffix = "-latlon",
  dstfile = paste0(sub("\\\\.[[:alnum:]]+$", "",
  as.character(object@file@name)), suffix, ".tiff"), of = "GTiff",
  r = "bilinear", t_srs = "+init=epsg:4326", dstalpha = T,
  dstnodata = "NA", overwrite = F, ...)
}
\arguments{
\item{object}{SAR object or character giving the address of a raster (e.g. GeoTiff).}

\item{suffix}{Character. This suffix is added to the image name in front of file extension.}

\item{dstfile}{Character. The destination file name.}

\item{...}{Arguments of \code{\link{gdalwarp}}.}
}
\value{
Character. The destination file name.
}
\description{
Projects SAR object using \code{\link{gdalwarp}}.
}
\examples{
# object@geolocationPoints <- spTransform(object@geolocationPoints, CRS(t_srs))
}

