% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarRecord.r
\docType{methods}
\name{sarRecord}
\alias{sarRecord}
\alias{sarRecord,RasterLayer-method}
\alias{sarRecord,character-method}
\alias{sarRecord,missing-method}
\title{Create a synthetic aperture radar (SAR) object}
\usage{
sarRecord(object, satellite = "", imgSubstring = "", polarization = NULL,
  ...)

\S4method{sarRecord}{RasterLayer}(object, satellite = "", imgSubstring = "",
  polarization = NULL, ...)

\S4method{sarRecord}{character}(object, satellite = "", imgSubstring = "",
  polarization = NULL, ...)

\S4method{sarRecord}{missing}(object, satellite = "", imgSubstring = "",
  polarization = NULL, ...)
}
\arguments{
\item{object}{either a character given the folder address of satellite data or a RasterLayer object.}

\item{satellite}{character. Either 'sentinel-1' or 'terrasar-x'.}

\item{imgSubstring}{character. This substring is searched for when image data is read.}

\item{polarization.}{character. E.g.'vv' or 'hh'. 
If the polarization is not given it will be determined from metadata.}
}
\value{
object of SAR class or subclass that is a well a RasterLayer object.
}
\description{
Creates a synthetic aperture radar (SAR) object. 
The input is either an existing raster or the folder address of satellite data.
}
\examples{
master <- sarRecord('extdata/S1A_IW_GRDH_1SDV_20151220T155517.SAFE/',
                    satellite = 'sentinel') 
slave <- sarRecord('extdata/S1A_IW_GRDH_1SDV_20151215T154711.SAFE/',
                    satellite = 'sentinel')
}

