% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.r
\docType{class}
\name{SARSet-class}
\alias{SARSet-class}
\title{SAR set class}
\description{
A class to handle a collection of synthetic aperture radar (SAR) records.
Objects of this class are lists.
}
\section{Slots}{

\describe{
\item{\code{crs}}{\code{\link{CRS}} object}
}}
\examples{
master <- sarRecord('extdata/S1A_IW_GRDH_1SDV_20151220T155517.SAFE/',
                    satellite = 'sentinel') 
slave <- sarRecord('extdata/S1A_IW_GRDH_1SDV_20151215T154711.SAFE/',
                    satellite = 'sentinel')

kiliSetAsc <- new('SARSet', c(master, slave))
              
alternatively
kiliSetAsc <- sarSet(c(master, slave))
}
\seealso{
\code{\link{SAR-class}}
}

