% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.r
\docType{class}
\name{SARSet-class}
\alias{SARSet-class}
\title{SAR set class}
\description{
A class to handle a collection of synthetic aperture radar (SAR) records.
Objects of this class are lists.
}
\section{Slots}{

\describe{
\item{\code{crs}}{\code{\link[sp]{CRS}} object}
}}
\examples{
# orbits 130 and 57
master <- new('Sentinel', address=system.file('extdata/S1A_IW_GRDH_1SDV_20151220T155517.SAFE', 
              package='ragram'))
slave <- new('Sentinel', address=system.file('extdata/S1A_IW_GRDH_1SDV_20151215T154711.SAFE', 
              package='ragram'))

# alternatively
master <- sarRecord(system.file('extdata/S1A_IW_GRDH_1SDV_20151220T155517.SAFE', package='ragram'),
                    satellite = 'sentinel') 
slave <- sarRecord(system.file('extdata/S1A_IW_GRDH_1SDV_20151215T154711.SAFE', package='ragram'),
                   satellite = 'sentinel')
                    
master@orbit
master@polarization
master@node

kiliSetAsc <- new('SARSet', c(master, slave))
              
# alternatively
kiliSetAsc <- sarSet(c(master, slave))
}
\seealso{
\code{\link{SAR-class}}
}

