% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.r
\docType{class}
\name{Sentinel-class}
\alias{Sentinel-class}
\title{Sentinel class}
\description{
A class to handle Sentinel-1 SAR data. 
This class inherits from the \code{\link{SAR-class}}.
}
\section{Slots}{

\describe{
\item{\code{address}}{Character. Address of the SAR data main folder.}

\item{\code{polarization}}{Character. Polarization of the SAR record (e.g. VV or HH).}

\item{\code{geolocationPoints}}{\code{\link{GeolocationPoints-class}} object that contains GCPs/TPs.}

\item{\code{node}}{Character. Direction of the satellite (ascending or descending).}

\item{\code{orbit}}{Integer. Orbit number (e.g. sentinel has 175 in total).}

\item{\code{imgSubstring}}{Character. This substring is searched for when image data is read.}
}}
\examples{
# orbits 130 and 57
master <- new('Sentinel', address='extdata/S1A_IW_GRDH_1SDV_20151220T155517.SAFE/')
slave <- new('Sentinel', address='extdata/S1A_IW_GRDH_1SDV_20151215T154711.SAFE/')

# alternatively
master <- sarRecord('extdata/S1A_IW_GRDH_1SDV_20151220T155517.SAFE/',
                    satellite = 'sentinel') 
slave <- sarRecord('extdata/S1A_IW_GRDH_1SDV_20151215T154711.SAFE/',
                    satellite = 'sentinel')
                    
master@orbit
master@polarization
master@node
}
\seealso{
\code{\link{SAR-class}}, \code{\link{GeolocationPoints-class}}, \code{\link{SARSet-class}}
}

