% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spTransform.r
\docType{methods}
\name{spTransform,GeolocationPoints,CRS-method}
\alias{spTransform,GeolocationPoints,CRS-method}
\title{spTransform for GeolocationPoints}
\usage{
\S4method{spTransform}{GeolocationPoints,CRS}(x, CRSobj, ...)
}
\arguments{
\item{x}{Object of the \code{\link{SAR-class}} or a subclass (e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{CRSobj}{Object of class \code{\link[sp]{CRS-class}}, or of class character in which case it is converted to \code{\link[sp]{CRS-class}}.}

\item{...}{Further arguments to \code{\link[sp]{spTransform}}.}
}
\description{
Transforms GeolocationPoints to coordinate reference system 
(see \code{\link[sp]{spTransform}}).
}
\seealso{
\code{\link{sarProject}}, \code{\link[sp]{spTransform}}, \code{\link[sp]{CRS-class}}
}

