% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zncc.r
\name{zncc}
\alias{zncc}
\title{Zero normalized cross correlation (ZNCC)}
\usage{
zncc(master, slave, u1, v1, u2, v2, n = 3, m = n)
}
\arguments{
\item{master}{\code{\link{SAR-class}} object (or a subclass e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{slave}{\code{\link{SAR-class}} object (or a subclass e.g. \code{\link{Sentinel-class}} or \code{\link{TSX-class}}).}

\item{u1}{Integer. Column, i.e. y coordinate of master pixel.}

\item{v1}{Integer. Row, i.e. x coordinate of master pixel.}

\item{u2}{Integer. Column, i.e. y coordinate of slave pixel.}

\item{v2}{Integer. Row, i.e. x coordinate of slave pixel.}

\item{n}{Integer. (Window size-1)/2 in x direction.}

\item{m}{Integer. (Window size-1)/2 in y direction.}
}
\description{
Calculates the zero normalized cross correlation (ZNCC) coefficient between two \code{\link{SAR-class}} objects.
This method also works for \code{\link{RasterLayer-class}} objects. The maximum ZNCC value determines disparity.
}

